pyproject.toml
[tool.poetry]
name = "pension-calculator"
version = "0.1.0"
description = ""
authors = ["Richard <richlyon@mac.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.5.0"
toml-config = "^0.1.3"
numpy-financial = "^1.0.0"
matplotlib = "3.5.2"
click = "^8.1.3"
py-moneyed = "^2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[tool.poetry.scripts]
compute = "pension_calculator.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
